Feat: Profile Popup Page System
