+++ Dungeon Jumak/Assets/00. Scripts/03. UI/CommonUI/Profile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f83189d7bacc2a64eb7c682867c3a433
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
